fix: keep lab capability planning out of main
