+++ src/main.rs
@@ -1,6 +1,6 @@
 use clap::{ArgMatches, Command, CommandFactory, FromArgMatches};
 use std::io::Read;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::process::Stdio;
 use std::time::Duration;
 
@@ -13,6 +13,9 @@ use homeboy::commands::cli;
 use homeboy::commands::utils::{args, entity_suggest, resource_policy, response as output};
 use homeboy::core::extension::load_all_extensions;
 
+use lab_offload_capabilities::lab_offload_source_path;
+
+mod lab_offload_capabilities;
 mod lab_offload_extension_parity;
 #[cfg(test)]
 mod reverse_lab_offload_tests;
@@ -709,8 +712,9 @@ fn run_lab_offload_inner(
         )
     })?;
     let source_path = lab_offload_source_path(normalized_args)?;
-    let capability_plan = lab_runner_capability_contract(command_kind, &source_path)
-        .map(homeboy::core::runner::lab_runner_capability_plan);
+    let capability_plan =
+        lab_offload_capabilities::lab_runner_capability_contract(command_kind, &source_path)
+            .map(homeboy::core::runner::lab_runner_capability_plan);
     if let Some(plan) = &capability_plan {
         let mode = match source {
             LabRunnerSelectionSource::Default => {
@@ -870,118 +874,6 @@ fn run_lab_offload_inner(
     Ok(Some(exit_code))
 }
 
-fn lab_runner_capability_contract(
-    command: &Commands,
-    source_path: &Path,
-) -> Option<homeboy::core::runner::LabRunnerCapabilityContract> {
-    if !command.supports_lab_runner() {
-        return None;
-    }
-
-    let command_label = match command {
-        Commands::Bench(args) if args.is_run_command() => "bench",
-        Commands::Refactor(args) if args.is_hot_resource_command() => "refactor",
-        Commands::Audit(args) if args.changed_since.is_none() && !args.conventions => "audit",
-        Commands::Lint(args) if args.is_full_workspace_run() => "lint",
-        Commands::Test(args) if args.changed_since.is_none() => "test",
-        Commands::Trace(_) => "trace",
-        _ => return None,
-    };
-
-    let mut required_tools = Vec::new();
-
-    if source_path.join("package.json").is_file() {
-        push_unique(
-            &mut required_tools,
-            homeboy::core::runner::RunnerRequiredTool::Node,
-        );
-        push_unique(
-            &mut required_tools,
-            homeboy::core::runner::RunnerRequiredTool::Npm,
-        );
-    }
-
-    if source_path.join("pnpm-lock.yaml").is_file() {
-        push_unique(
-            &mut required_tools,
-            homeboy::core::runner::RunnerRequiredTool::Node,
-        );
-        push_unique(
-            &mut required_tools,
-            homeboy::core::runner::RunnerRequiredTool::Pnpm,
-        );
-    }
-
-    if source_path.join("composer.json").is_file() {
-        push_unique(
-            &mut required_tools,
-            homeboy::core::runner::RunnerRequiredTool::Php,
-        );
-        push_unique(
-            &mut required_tools,
-            homeboy::core::runner::RunnerRequiredTool::Composer,
-        );
-    }
-
-    if has_docker_signal(source_path) {
-        push_unique(
-            &mut required_tools,
-            homeboy::core::runner::RunnerRequiredTool::Docker,
-        );
-    }
-
-    Some(homeboy::core::runner::LabRunnerCapabilityContract {
-        command: command_label,
-        required_tools,
-        requires_playwright: matches!(command, Commands::Trace(_)),
-    })
-}
-
-fn push_unique<T: PartialEq>(items: &mut Vec<T>, item: T) {
-    if !items.contains(&item) {
-        items.push(item);
-    }
-}
-
-fn has_docker_signal(source_path: &Path) -> bool {
-    [
-        "Dockerfile",
-        "docker-compose.yml",
-        "docker-compose.yaml",
-        "compose.yml",
-        "compose.yaml",
-    ]
-    .iter()
-    .any(|name| source_path.join(name).is_file())
-}
-
-fn lab_offload_source_path(args: &[String]) -> homeboy::core::Result<PathBuf> {
-    let mut iter = args.iter().skip(1).peekable();
-    while let Some(arg) = iter.next() {
-        if arg == "--" {
-            break;
-        }
-        if arg == "--path" {
-            let value = iter.next().ok_or_else(|| {
-                homeboy::core::Error::validation_invalid_argument(
-                    "path",
-                    "--path requires a value before Lab offload can sync the workspace",
-                    None,
-                    None,
-                )
-            })?;
-            return Ok(PathBuf::from(shellexpand::tilde(value).to_string()));
-        }
-        if let Some(value) = arg.strip_prefix("--path=") {
-            return Ok(PathBuf::from(shellexpand::tilde(value).to_string()));
-        }
-    }
-
-    std::env::current_dir().map_err(|err| {
-        homeboy::core::Error::internal_io(err.to_string(), Some("read cwd".to_string()))
-    })
-}
-
 fn rewrite_lab_offload_args(args: &[String], remote_path: &str) -> Vec<String> {
     let mut stripped = Vec::with_capacity(args.len());
     let mut iter = args.iter().peekable();
@@ -1125,10 +1017,9 @@ fn extract_parent_command_from_error(e: &clap::Error) -> Option<String> {
 #[cfg(test)]
 mod tests {
     use super::{
-        lab_offload_source_path, lab_runner_capability_contract,
-        prepare_lab_runner_for_offload_with, resolve_lab_runner_selection_from_default,
-        rewrite_lab_offload_args, LabRunnerPreparation, LabRunnerSelection,
-        LabRunnerSelectionSource,
+        lab_offload_source_path, prepare_lab_runner_for_offload_with,
+        resolve_lab_runner_selection_from_default, rewrite_lab_offload_args, LabRunnerPreparation,
+        LabRunnerSelection, LabRunnerSelectionSource,
     };
     use clap::Parser;
     use homeboy::cli_surface::Commands;
@@ -1157,34 +1048,6 @@ mod tests {
             args: Vec::new(),
             json_summary: false,
         }
-    }
-
-    #[test]
-    fn lab_runner_capability_contract_detects_workspace_tool_signals() {
-        let dir = tempfile::tempdir().expect("temp project");
-        std::fs::write(dir.path().join("package.json"), "{}").expect("package.json");
-        std::fs::write(dir.path().join("pnpm-lock.yaml"), "lockfileVersion: '9'\n")
-            .expect("pnpm-lock");
-        std::fs::write(dir.path().join("composer.json"), "{}").expect("composer.json");
-        std::fs::write(dir.path().join("Dockerfile"), "FROM scratch\n").expect("Dockerfile");
-        let command = homeboy::cli_surface::Cli::try_parse_from(["homeboy", "lint"])
-            .expect("parse")
-            .command;
-
-        let contract = lab_runner_capability_contract(&command, dir.path()).expect("contract");
-
-        assert_eq!(contract.command, "lint");
-        assert_eq!(
-            contract.required_tools,
-            vec![
-                homeboy::core::runner::RunnerRequiredTool::Node,
-                homeboy::core::runner::RunnerRequiredTool::Npm,
-                homeboy::core::runner::RunnerRequiredTool::Pnpm,
-                homeboy::core::runner::RunnerRequiredTool::Php,
-                homeboy::core::runner::RunnerRequiredTool::Composer,
-                homeboy::core::runner::RunnerRequiredTool::Docker,
-            ]
-        );
     }
 
     #[test]
